Add files via upload
I recently completed the task of developing an HTML and CSS website. This project involved creating a visually appealing and functional web page using HTML for the structure and CSS for the styling. I ensured that the design was clean, modern, and user-friendly. Additionally, I focused on making the website responsive, meaning it can adapt to various screen sizes and devices, providing an optimal viewing experience for users on desktops, tablets, and smartphones.